Extend TypedPDA with optional state mapping (merged from StateTypedPDA)
